Feature: Add video gallery component and integrate with existing gallery structure
