+++ package.json
@@ -1,63 +1,64 @@
 {
-  "name": "boda-nc",
-  "type": "module",
-  "version": "1.0.0",
-  "scripts": {
-    "dev": "astro dev",
-    "start": "astro dev",
-    "build": "astro check && astro build",
-    "preview": "astro preview",
-    "astro": "astro",
-    "format": "prettier -w \"**/*\" --ignore-unknown --cache",
-    "script:create-meta": "pnpx tsx scripts/create-meta-from-gallery.ts",
-    "script:create-thumb": "pnpx tsx scripts/create-thumbnails-gallery.ts",
-    "script:conver-images": "pnpx tsx scripts/convert-gallery-images.ts",
-    "script:conver-images-remove": "pnpx tsx scripts/convert-gallery-images.ts rm",
-    "script:rename-images": "pnpx tsx scripts/rename-gallery-images.ts",
-    "script:prepare-deploy-gallery": "pnpx tsx scripts/prepare-deploy-gallery.ts"
-  },
-  "dependencies": {
-    "@appnest/masonry-layout": "2.2.3",
-    "@astrojs/mdx": "4.0.8",
-    "@astrojs/react": "4.2.1",
-    "@astrojs/rss": "4.0.11",
-    "@astrojs/sitemap": "3.2.1",
-    "@astrojs/vercel": "8.1.1",
-    "@fontsource-variable/playfair-display": "5.1.0",
-    "@fontsource-variable/raleway": "5.1.0",
-    "@fontsource-variable/work-sans": "5.1.0",
-    "@fontsource/cinzel-decorative": "5.1.0",
-    "@iconify-json/tabler": "1.2.16",
-    "@react-email/components": "0.0.33",
-    "@react-email/render": "1.0.5",
-    "@tailwindcss/forms": "0.5.9",
-    "@tailwindcss/vite": "4.0.6",
-    "@types/js-cookie": "3.0.6",
-    "@types/react": "19.0.10",
-    "@types/react-dom": "19.0.4",
-    "astro": "5.3.0",
-    "astro-auto-import": "0.4.4",
-    "astro-icon": "1.1.4",
-    "astro-seo": "0.8.4",
-    "glob": "11.0.1",
-    "image-meta": "0.2.1",
-    "js-cookie": "3.0.5",
-    "photoswipe": "5.4.4",
-    "react": "19.0.0",
-    "react-dom": "19.0.0",
-    "resend": "4.1.2",
-    "sonner": "2.0.1",
-    "tailwindcss": "4.0.6",
-    "typescript": "5.8.3"
-  },
-  "devDependencies": {
-    "@playform/compress": "0.1.7",
-    "prettier-plugin-astro": "0.14.1",
-    "prettier-plugin-tailwindcss": "0.6.11",
-    "sharp": "0.33.0",
-    "ts-node": "10.9.2"
-  },
-  "overrides": {
-    "sharp": "0.33.0"
-  }
+	"name": "boda-nc",
+	"type": "module",
+	"version": "1.0.0",
+	"scripts": {
+		"dev": "astro dev",
+		"start": "astro dev",
+		"build": "astro check && astro build",
+		"preview": "astro preview",
+		"astro": "astro",
+		"format": "prettier -w \"**/*\" --ignore-unknown --cache",
+		"script:create-meta": "pnpx tsx scripts/create-meta-from-gallery.ts",
+		"script:create-thumb": "pnpx tsx scripts/create-thumbnails-gallery.ts",
+		"script:conver-images": "pnpx tsx scripts/convert-gallery-images.ts",
+		"script:conver-images-remove": "pnpx tsx scripts/convert-gallery-images.ts rm",
+		"script:rename-images": "pnpx tsx scripts/rename-gallery-images.ts",
+		"script:prepare-deploy-gallery": "pnpx tsx scripts/prepare-deploy-gallery.ts"
+	},
+	"dependencies": {
+		"@appnest/masonry-layout": "2.2.3",
+		"@astrojs/mdx": "4.0.8",
+		"@astrojs/react": "4.2.1",
+		"@astrojs/rss": "4.0.11",
+		"@astrojs/sitemap": "3.2.1",
+		"@astrojs/vercel": "8.1.1",
+		"@fontsource-variable/playfair-display": "5.1.0",
+		"@fontsource-variable/raleway": "5.1.0",
+		"@fontsource-variable/work-sans": "5.1.0",
+		"@fontsource/cinzel-decorative": "5.1.0",
+		"@iconify-json/tabler": "1.2.16",
+		"@justinribeiro/lite-youtube": "1.8.2",
+		"@react-email/components": "0.0.33",
+		"@react-email/render": "1.0.5",
+		"@tailwindcss/forms": "0.5.9",
+		"@tailwindcss/vite": "4.0.6",
+		"@types/js-cookie": "3.0.6",
+		"@types/react": "19.0.10",
+		"@types/react-dom": "19.0.4",
+		"astro": "5.3.0",
+		"astro-auto-import": "0.4.4",
+		"astro-icon": "1.1.4",
+		"astro-seo": "0.8.4",
+		"glob": "11.0.1",
+		"image-meta": "0.2.1",
+		"js-cookie": "3.0.5",
+		"photoswipe": "5.4.4",
+		"react": "19.0.0",
+		"react-dom": "19.0.0",
+		"resend": "4.1.2",
+		"sonner": "2.0.1",
+		"tailwindcss": "4.0.6",
+		"typescript": "5.8.3"
+	},
+	"devDependencies": {
+		"@playform/compress": "0.1.7",
+		"prettier-plugin-astro": "0.14.1",
+		"prettier-plugin-tailwindcss": "0.6.11",
+		"sharp": "0.33.0",
+		"ts-node": "10.9.2"
+	},
+	"overrides": {
+		"sharp": "0.33.0"
+	}
 }
