+++ src/components/Gallery/ButtonToTop.tsx
@@ -1,3 +1,5 @@
+import { useEffect, useState } from "react";
+
 export default function ButtonToTop({
 	className = "",
 	scrollTo = "#gallery",
@@ -5,7 +7,16 @@ export default function ButtonToTop({
 	className?: string;
 	scrollTo?: string;
 }) {
-	return (
+	const [showButtonToTop, setShowButtonToTop] = useState(false);
+	useEffect(() => {
+		const handleScroll = () => {
+			setShowButtonToTop(window.scrollY > window.innerHeight);
+		};
+		window.addEventListener("scroll", handleScroll);
+		return () => window.removeEventListener("scroll", handleScroll);
+	}, []);
+
+	return showButtonToTop ? (
 		<a
 			href={scrollTo}
 			className={`fixed right-4 bottom-4 z-50 flex h-12 w-12 items-center justify-center rounded-full bg-white text-gray-800 shadow-lg transition-transform hover:scale-110 focus:ring-2 focus:ring-gray-500 focus:outline-none ${className}`}
@@ -21,5 +32,7 @@ export default function ButtonToTop({
 				/>
 			</svg>
 		</a>
+	) : (
+		<></>
 	);
 }
